requirements.txt
Django==2.1.2
django-appconf==1.0.2
django-bootstrap-admin==0.3.4
django-colorfield==0.1.14
django-countries==5.0
django-encrypted-model-fields==0.5.3
django-import-export==0.5.1
django-jet==1.0.7
django-nested-inline==0.3.7
django-premailer==0.2.0
django-sendgrid-v5==0.6.87
django-smart-selects==1.5.2
django-storages==1.6.5
django-suit==0.2.26
django-timezone-field==2.0
flickrapi==2.4.0
Pillow==5.3.0
pycodestyle==2.4.0
psycopg2==2.7.3.2
requests==2.18.4
six==1.11.0
spotipy==2.4.4

